Refactor Meeting Date to DateTime
