+++ src/main/java/staffconnect/model/meeting/comparator/MeetingDateThenDescriptionComparator.java
@@ -1,6 +1,6 @@
 package staffconnect.model.meeting.comparator;
 
-import static staffconnect.model.meeting.comparator.MeetingDateComparator.MEETING_DATE_COMPARATOR;
+import static staffconnect.model.meeting.comparator.MeetingDateTimeComparator.MEETING_DATE_COMPARATOR;
 import static staffconnect.model.meeting.comparator.MeetingDescriptionComparator.MEETING_DESCRIPTION_COMPARATOR;
 
 import java.util.Comparator;
