+++ src/main/java/staffconnect/model/person/comparators/MeetingListComparator.java
@@ -1,6 +1,6 @@
 package staffconnect.model.person.comparators;
 
-import static staffconnect.model.meeting.comparator.MeetingDateComparator.MEETING_DATE_COMPARATOR;
+import static staffconnect.model.meeting.comparator.MeetingDateTimeComparator.MEETING_DATE_COMPARATOR;
 
 import java.util.Comparator;
